// cypress/support/types/recruitmentFormData.d.ts
export interface IRecruitmentFormData {
  vacancyName: string
  vacancyStatus: boolean
  vacancyPublished: boolean
  jobTitleName: string
  jobDescription: string
  jobNote: string
  interviewTitle: string
  interviewerNameHint: string
  interviewDate: string
  interviewTime: string
  candidateFirstName: string
  candidateLastName: string
  candidateEmail: string
  candidateStatus: string
}